local_api/requirements.txt
flask
tabula-py
JPype1